rename filenames to reflect the meaning of the HTML elements
