+++ packages/@atjson/source-html/src/annotations/heading-1.ts
@@ -1,4 +1,4 @@
-import { BlockAnnotation } from '@atjson/document';
+import { BlockAnnotation } from '../../../document/dist/commonjs';
 
 export default class Heading1 extends BlockAnnotation {
   static vendorPrefix = 'html';
