+++ packages/@atjson/source-html/src/annotations/index.ts
@@ -1,26 +1,26 @@
-export { default as Anchor } from './a';
-export { default as Bold } from './b';
+export { default as Anchor } from './anchor';
+export { default as Bold } from './bold';
 export { default as Blockquote } from './blockquote';
-export { default as Break } from './br';
+export { default as Break } from './break';
 export { default as Code } from './code';
-export { default as DeletedText } from './del';
-export { default as Emphasis } from './em';
-export { default as Heading1 } from './h1';
-export { default as Heading2 } from './h2';
-export { default as Heading3 } from './h3';
-export { default as Heading4 } from './h4';
-export { default as Heading5 } from './h5';
-export { default as Heading6 } from './h6';
-export { default as HorizontalRule } from './hr';
-export { default as Image } from './img';
-export { default as Italic } from './i';
-export { default as ListItem } from './li';
-export { default as OrderedList } from './ol';
-export { default as Paragraph } from './p';
-export { default as PreformattedText } from './pre';
-export { default as Strikethrough } from './s';
+export { default as DeletedText } from './deleted-text';
+export { default as Emphasis } from './emphasis';
+export { default as Heading1 } from './heading-1';
+export { default as Heading2 } from './heading-2';
+export { default as Heading3 } from './heading-3';
+export { default as Heading4 } from './heading-4';
+export { default as Heading5 } from './heading-5';
+export { default as Heading6 } from './heading-6';
+export { default as HorizontalRule } from './horizontal-rule';
+export { default as Image } from './image';
+export { default as Italic } from './italic';
+export { default as ListItem } from './list-item';
+export { default as OrderedList } from './ordered-list';
+export { default as Paragraph } from './paragraph';
+export { default as PreformattedText } from './preformatted-text';
+export { default as Strikethrough } from './strikethrough';
 export { default as Strong } from './strong';
-export { default as Subscript } from './sub';
-export { default as Superscript } from './sup';
-export { default as UnorderedList } from './ul';
-export { default as Underline } from './u';
+export { default as Subscript } from './subscript';
+export { default as Superscript } from './superscript';
+export { default as UnorderedList } from './unordered-list';
+export { default as Underline } from './underline';
