+++ packages/@atjson/source-html/src/annotations/list-item.ts
@@ -1,4 +1,4 @@
-import { BlockAnnotation } from '@atjson/document';
+import { BlockAnnotation } from '../../../document/dist/commonjs';
 
 export default class ListItem extends BlockAnnotation {
   static vendorPrefix = 'html';
