+++ packages/@atjson/source-html/src/annotations/ordered-list.ts
@@ -1,4 +1,4 @@
-import { BlockAnnotation } from '@atjson/document';
+import { BlockAnnotation } from '../../../document/dist/commonjs';
 
 export default class OrderedList extends BlockAnnotation {
   static vendorPrefix = 'html';
