+++ packages/@atjson/source-html/src/annotations/paragraph.ts
@@ -1,4 +1,4 @@
-import { BlockAnnotation } from '@atjson/document';
+import { BlockAnnotation } from '../../../document/dist/commonjs';
 
 export default class Paragraph extends BlockAnnotation {
   static vendorPrefix = 'html';
