+++ packages/@atjson/source-html/src/annotations/subscript.ts
@@ -1,4 +1,4 @@
-import { InlineAnnotation } from '@atjson/document';
+import { InlineAnnotation } from '../../../document/dist/commonjs';
 
 export default class Subscript extends InlineAnnotation {
   static vendorPrefix = 'html';
