+++ packages/@atjson/source-html/src/annotations/superscript.ts
@@ -1,4 +1,4 @@
-import { InlineAnnotation } from '@atjson/document';
+import { InlineAnnotation } from '../../../document/dist/commonjs';
 
 export default class Superscript extends InlineAnnotation {
   static vendorPrefix = 'html';
