+++ packages/@atjson/source-html/src/annotations/unordered-list.ts
@@ -1,4 +1,4 @@
-import { BlockAnnotation } from '@atjson/document';
+import { BlockAnnotation } from '../../../document/dist/commonjs';
 
 export default class UnorderedList extends BlockAnnotation {
   static vendorPrefix = 'html';
